Restructured command folders
