+++ src/main/java/frc/robot/RobotContainer.java
@@ -16,15 +16,12 @@ import edu.wpi.first.wpilibj.controller.RamseteController;
 import edu.wpi.first.wpilibj.controller.SimpleMotorFeedforward;
 import edu.wpi.first.wpilibj.geometry.Pose2d;
 import edu.wpi.first.wpilibj.geometry.Rotation2d;
-import edu.wpi.first.wpilibj.geometry.Translation2d;
 import edu.wpi.first.wpilibj.kinematics.ChassisSpeeds;
 import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.trajectory.Trajectory;
 import edu.wpi.first.wpilibj.trajectory.TrajectoryConfig;
 import edu.wpi.first.wpilibj.trajectory.TrajectoryGenerator;
-import edu.wpi.first.wpilibj.trajectory.constraint.CentripetalAccelerationConstraint;
-import edu.wpi.first.wpilibj.trajectory.constraint.DifferentialDriveKinematicsConstraint;
 import edu.wpi.first.wpilibj.trajectory.constraint.DifferentialDriveVoltageConstraint;
 import edu.wpi.first.wpilibj.util.Units;
 import edu.wpi.first.wpilibj2.command.Command;
@@ -43,12 +40,12 @@ import frc.robot.subsystems.BallLauncher;
 import frc.robot.subsystems.ClimberSubsystem;
 import frc.robot.subsystems.LightsSubsystem;
 
-import frc.robot.commands.ManualHopper;
-import frc.robot.commands.SetSolidColor;
+import frc.robot.commands.hopper.ManualHopper;
+import frc.robot.commands.light.SetSolidColor;
 import frc.robot.Constants.AutoConstants;
 import frc.robot.Constants.DriveConstants;
-import frc.robot.commands.BasicCommandGroup;
-import frc.robot.commands.SimpleAutoGroup;
+import frc.robot.commands.auto.BasicCommandGroup;
+import frc.robot.commands.auto.SimpleAutoGroup;
 import frc.robot.commands.drive.FieldOrientedTurn;
 import frc.robot.commands.auto.Center8BallAuto;
