+++ src/main/java/frc/robot/commands/auto/BasicCommandGroup.java
@@ -5,12 +5,12 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.auto;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.commands.DriveToWall;
-import frc.robot.commands.ShootBallsGroup;
-import frc.robot.commands.HopperShift;
+import frc.robot.commands.drive.DriveToWall;
+import frc.robot.commands.auto.ShootBallsGroup;
+import frc.robot.commands.hopper.HopperShift;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.BallLauncher;
 import frc.robot.subsystems.FeederSubsystem;
