+++ src/main/java/frc/robot/commands/auto/SimpleAutoGroup.java
@@ -5,10 +5,10 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.auto;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.commands.DriveTimed;
+import frc.robot.commands.drive.DriveTimed;
 import frc.robot.subsystems.DriveSubsystem;
 
 public class SimpleAutoGroup extends SequentialCommandGroup {
