+++ src/main/java/frc/robot/commands/light/SetSolidColor.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.light;
 
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
